+++ packages/node/src/behavior/system/network/NetworkRuntime.ts
@@ -49,6 +49,7 @@ export abstract class NetworkRuntime {
         } finally {
             this.#owner.behaviors.internalsOf(NetworkBehavior).runtime = undefined;
         }
+        await this.owner.prepareRuntimeShutdown();
         await this.#owner.act(agent => this.owner.lifecycle.offline.emit(agent.context));
     }