+++ packages/node/src/behaviors/general-diagnostics/GeneralDiagnosticsServer.ts
@@ -65,10 +65,8 @@ export class GeneralDiagnosticsServer extends Base {
         }
 
         const lifecycle = this.endpoint.lifecycle as NodeLifecycle;
-
-        if (lifecycle.online !== undefined) {
-            this.reactTo(lifecycle.online, this.#online, { lock: true });
-        }
+        this.reactTo(lifecycle.online, this.#online, { lock: true });
+        this.reactTo(lifecycle.goingOffline, this.#goingOffline, { lock: true });
 
         if (this.events.activeHardwareFaults$Changed !== undefined) {
             this.reactTo(this.events.activeHardwareFaults$Changed, this.#triggerActiveHardwareFaultsChangedEvent);
@@ -294,6 +292,11 @@ export class GeneralDiagnosticsServer extends Base {
         await this.#updateNetworkList();
     }
 
+    #goingOffline() {
+        this.internal.lastTotalOperationalHoursTimer?.stop();
+        this.#updateTotalOperationalHoursCounter();
+    }
+
     #updateTotalOperationalHoursCounter() {
         const now = Time.nowMs();
         const elapsedTime = now - this.internal.lastTotalOperationalHoursCounterUpdateTime;
@@ -360,12 +363,6 @@ export class GeneralDiagnosticsServer extends Base {
                 iPv6Addresses: ipV6.slice(0, 8).map(ip => ipv4ToBytes(ip)),
                 type: type ?? networkType,
             }));
-    }
-
-    override async [Symbol.asyncDispose]() {
-        this.internal.lastTotalOperationalHoursTimer?.stop();
-        this.#updateTotalOperationalHoursCounter();
-        await super[Symbol.asyncDispose]?.();
     }
 }